changed: extracted common code to query decorator
